Add TLS passthrough, and metrics collection
Add an option for TLS passthrough for cases where backend services do
their own end-to-end encryption. In this mode, the proxy can't do client
authentication.

Collect some metrics for each backend service: number of connections,
number of bytes sent and received.

Optionally make these metrics available on an admin console, which is a
special backend service with `Mode: CONSOLE`. Connections to this
backend are directed to an internal http server. The metrics will
eventually be exported to prometheus (or something like that).

Add a flag to run the proxy with an ephemeral certificate manager
instead of letsencrypt (`--use-ephemeral-certificate-manager`). This is
useful for testing.