Implement a local OIDC provider
Implement a local OIDC provider _server_ to enable SSO on backend
services that support OIDC. This new built-in server uses the existing
user identity provided by SAML and/or OIDC.

Claims can be rewritten to match the needs of the local clients, e.g.
the `email` claim can used to derive a `preferred_username` claim by
removing the @domain part.