Passkey ID refresh (#38)
### Description

Add an option to refresh the identity used with passkeys.

Passkey SSO uses another identity provider to authenticate the user
before their passkey is registered. With this new option, tlsproxy can
be configured to re-check the user's identity with the other identity
provider after the interval specified in `refreshInterval`. By default,
the re-check is not enabled.

### Type of change

* [ ] New feature
* [x] Feature improvement
* [ ] Bug fix
* [ ] Documentation
* [ ] Cleanup / refactoring
* [ ] Other (please explain)


### How is this change tested ?

* [ ] Unit tests
* [x] Manual tests (explain)
* [ ] Tests are not needed