Check the golang image
Make sure that the golang image we use to build the docker image is
built for the right platform.

This is intended to guard against the case where the base image is built
for linux/386, but not for linux/amd64 (yet?). If/when that happens, the
docker image build job will fail.