Expand passkeys to work without resident keys (#159)
### Description

Expand passkeys to work without resident keys (aka discoverable
credentials).

This change adds a UI component to the passkey login screen to capture
the user's email address. When the email address is provided, the user
can log in with any passkeys they have registered. If the email address
is not provided, only passkeys with resident keys can be used.

### Type of change

* [ ] New feature
* [x] Feature improvement
* [ ] Bug fix
* [ ] Documentation
* [ ] Cleanup / refactoring
* [ ] Other (please explain)

### How is this change tested ?

* [x] Unit tests
* [x] Manual tests (explain)
* [ ] Tests are not needed